breaking down CNM to CNM, CNMHelper and ScaleUpCharging
